Made a couple of dependency refactorings / structure updates:
Based on this blogpost http://csausdev.wordpress.com/2010/12/17/dependency-injection-in-node-js/,
I changed it so that the Thread controller no longer creates its own connection - this now happens in app.js, and the connection is passed to the thread controller, which in turn passes it to the thread and post models.

The thread controller now also sets his own routes, instead of the routes.js module. It's also refactored a bit so its exports are neater - this should make it easier to test. I should check if one really has to export something in order to make it testable though.