Added a proper unit test for the Thread controller.
Refactored the thread controller so that instead of getting a connection and creates his own Thread and Post model, it gets those from the outside
, making them mockable. The new thread controller unit test heavily depends on mocked thread / post models and requests / responses.

Also updated the runTests method to run this test. Integration tests still have to be run manually, and I think that's the way it should remain fo
r now.